--- sonar-tsql-plugin/src/main/java/org/sonar/plugins/tsql/sensors/custom/ILinesProvider.java
@@ -1,5 +1,0 @@
-package org.sonar.plugins.tsql.sensors.custom;
-
-public interface ILinesProvider {
-	int getLine(ParsedNode node);
-}
