+++ sonar-tsql-plugin/src/main/java/org/sonar/plugins/tsql/sensors/custom/lines/ILinesProvider.java
@@ -1,0 +1,7 @@
+package org.sonar.plugins.tsql.sensors.custom.lines;
+
+import org.sonar.plugins.tsql.sensors.custom.ParsedNode;
+
+public interface ILinesProvider {
+	int getLine(ParsedNode node);
+}
