+++ sonar-tsql-plugin/src/main/java/org/sonar/plugins/tsql/sensors/custom/nodes/SiblingsNodesProvider.java
@@ -1,4 +1,4 @@
-package org.sonar.plugins.tsql.sensors.custom.providers;
+package org.sonar.plugins.tsql.sensors.custom.nodes;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,27 +6,24 @@ import java.util.List;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.sonar.plugins.tsql.sensors.custom.ParsedNode;
 
-public class ParentNodesProvider implements INodesProvider {
+public class SiblingsNodesProvider implements INodesProvider {
 
-	@Override
 	public ParsedNode[] getNodes(final ParsedNode node) {
 		final List<ParsedNode> nodes = new ArrayList<ParsedNode>();
 
 		if (node == null || node.getItem() == null) {
 			return nodes.toArray(new ParsedNode[0]);
 		}
-
 		final ParseTree item = node.getItem();
 
-		ParseTree parent = item.getParent();
+		final ParseTree parent = item.getParent();
 		if (parent == null) {
 			return new ParsedNode[0];
 		}
-		int d = 0;
-		while (parent != null) {
-			d--;
-			nodes.add(new ParsedNode(parent, node.getRule(), node.getRepository(), d));
-			parent = parent.getParent();
+		final int c = parent.getChildCount();
+		for (int i = 0; i < c; i++) {
+			final ParseTree child = parent.getChild(i);
+			nodes.add(new ParsedNode(child, node.getRule(), node.getRepository(), i));
 		}
 
 		return nodes.toArray(new ParsedNode[0]);
