+++ sonar-tsql-plugin/src/test/java/org/sonar/plugins/tsql/sensors/custom/providers/SiblingNodesProviderTest.java
@@ -6,6 +6,7 @@ import org.junit.Test;
 import org.sonar.plugins.tsql.helpers.Antlr4Utils;
 import org.sonar.plugins.tsql.rules.custom.Rule;
 import org.sonar.plugins.tsql.sensors.custom.ParsedNode;
+import org.sonar.plugins.tsql.sensors.custom.nodes.SiblingsNodesProvider;
 
 public class SiblingNodesProviderTest {
